--- ISPprojektas/app/Http/Requests/CommentController.php
@@ -1,36 +1,0 @@
-<?php
-
-namespace App\Http\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class CommentRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'rating' => 'required'
-        ];
-    }
-
-    public function messages(){
-      return[
-        'rating.required' => 'Ivertinimas yra privalomas'
-      ];
-    }
-}
